requirements.txt
# --- Core Framework ---
Flask>=3.1.0

# --- Database & Migrations ---
SQLAlchemy>=2.0.46
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.1.0
alembic>=1.14.0

# --- Database Drivers ---
psycopg2-binary>=2.9.10
psycopg[binary]>=3.2.0

# --- Authentication & Firebase ---
firebase-admin>=6.6.0
PyJWT[crypto]>=2.10.0

# --- Utilities ---
python-dotenv>=1.0.1
requests>=2.32.0
cryptography>=42.0.0

# --- CORS (ADD THIS) ---
Flask-Cors>=4.0.0

# --- Production ---
gunicorn>=23.0.0
